+++ test/peddler/api/orders_v0_test.rb
@@ -3,6 +3,7 @@
 require "helper"
 
 require "peddler/api/orders_v0"
+require "peddler/marketplace"
 
 module Peddler
   class API
@@ -10,7 +11,7 @@ module Peddler
       include FeatureHelpers
 
       def test_get_orders
-        res = api.sandbox.get_orders(["ATVPDKIKX0DER"], created_after: "TEST_CASE_200")
+        res = api.sandbox.get_orders(Marketplace.ids("US"), created_after: "TEST_CASE_200")
 
         assert_predicate(res.status, :ok?)
       end