+++ test/peddler/api/product_type_definitions_2020_09_01_test.rb
@@ -12,7 +12,7 @@ module Peddler
       def test_get_definitions_product_type
         res = api.get_definitions_product_type(
           "PRODUCT",
-          "A1F83G8C2ARO7P",
+          Marketplace.id("UK"),
           requirements: "LISTING_OFFER_ONLY",
         )